playwright youtube videos links extractor
This script is a work in progress. It extracts youtube links of youtube videos from a particular channel.
The user will have to provide links to the youtube channel; videos tab and the number of videos to be extracted.
Links are relative and need to be converted to absolute URLs.